fix(tests): move DB-backed search tests to tests/e2e (codex P1)
The autouse `_clean_search_tables` fixture in tests/search/conftest.py
opened a session for EVERY test under that directory, which broke unit-tier
CI collection (no Postgres service). Move the two DB-touching test files
(test_worker, test_service) into tests/e2e/, fold their fixtures into
tests/e2e/conftest.py (scoped to the seed fixtures rather than autouse so
non-search e2e tests don't pay per-test TRUNCATE), and drop the old
search conftest. The remaining tests/search/ files (chunker, RRF, snippet,
text_extract, lexical_backends, embedding_provider) are pure and import
cleanly.

Addresses codex review comment 3399849914.
